Revert renaming sequential-pane-display.
